Complete SOLID refactor phases 2–3 and fix quality checks.
Introduce CompiledFields on FormatParser, split core regex and PyO3
format_parser/findall modules, route convert_value through builtin_convert,
and clear ruff/clippy/mypy/rustfmt issues.
