+++ formatparse-core/src/types/regex/mod.rs
@@ -1,14 +1,16 @@
 //! Regex generation from field specifications.
 
+mod custom;
+mod datetime_fragment;
 mod helpers;
+mod numeric;
+mod string;
 #[cfg(test)]
 mod tests;
 
 pub use helpers::strftime_to_regex;
 
 use crate::types::definitions::{FieldSpec, FieldType};
-use helpers::{int_width_precision_bounds, wrap_field_lookbehind, STRING_WIDTH_PRECISION_CHAR};
-use regex;
 use std::collections::HashMap;
 
 impl FieldSpec {
@@ -21,432 +23,34 @@ impl FieldSpec {
         next_field_is_greedy: Option<bool>,
         allow_empty_delimited: bool,
     ) -> String {
-        let fmt_char = STRING_WIDTH_PRECISION_CHAR;
-        let base_pattern = match &self.field_type {
-            FieldType::String => {
-                // Handle alignment and width for strings
-                if let Some(prec) = self.precision {
-                    // Precision specified: match exactly 'precision' characters
-                    // If alignment is also specified, allow fill characters in appropriate positions
-                    if let Some(align) = self.alignment {
-                        let fill_ch = self.fill.unwrap_or(' ');
-                        let fill_escaped = regex::escape(&fill_ch.to_string());
-                        // When width == precision, the formatted field is exactly `prec` characters
-                        // wide; optional fill segments would let the capture extend past that width
-                        // and consume text meant for a following literal (GitHub issue #97).
-                        let width_equals_precision = self.width == Some(prec);
-                        match align {
-                            '<' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", fmt_char, prec)
-                                } else {
-                                    // Left-aligned: content (precision chars) + optional trailing fill
-                                    format!("{}{{{}}}(?:{}*)", fmt_char, prec, fill_escaped)
-                                }
-                            }
-                            '>' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", fmt_char, prec)
-                                } else {
-                                    // Right-aligned: optional leading fill + exactly `prec` characters.
-                                    // Leading fill must be non-greedy so we do not consume the entire slice
-                                    // before `.{{prec}}` when another field follows (issue #88 / parse#218).
-                                    format!("(?:{}*?){}{{{}}}", fill_escaped, fmt_char, prec)
-                                }
-                            }
-                            '^' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", fmt_char, prec)
-                                } else {
-                                    // Center-aligned: optional leading fill + content + optional trailing fill.
-                                    // Non-greedy leading fill for the same boundary reason as `>`.
-                                    format!(
-                                        "(?:{}*?){}{{{}}}(?:{}*)",
-                                        fill_escaped, fmt_char, prec, fill_escaped
-                                    )
-                                }
-                            }
-                            _ => format!("{}{{{}}}", fmt_char, prec),
-                        }
-                    } else {
-                        // Precision only, no alignment: match exactly 'precision' characters
-                        format!("{}{{{}}}", fmt_char, prec)
-                    }
-                } else if let Some(width) = self.width {
-                    // Width only (no precision):
-                    // - If there's a next field with precision (like {:.4}), use greedy (at least width)
-                    // - If there's a next field without precision (like {}), use exact width
-                    // - If it's the last field, use greedy (at least width)
-                    match next_field_is_greedy {
-                        Some(false) => format!("{}{{{}}}", fmt_char, width), // Exact when followed by non-greedy field
-                        _ => format!("{}{{{},}}", fmt_char, width), // Greedy when followed by greedy field or last field
-                    }
-                } else if self.alignment.is_some() {
-                    // Alignment specified but no width - match with optional surrounding whitespace
-                    // For alignment, we want to capture only the text value (without padding spaces)
-                    // The padding spaces are part of the alignment formatting, not the value
-                    match self.alignment {
-                        // Left: capture text, then allow trailing spaces (non-capturing)
-                        Some('<') => r"([^\{\}\s]+(?:\s+[^\{\}\s]+)*?)(?:\s*)".to_string(),
-                        // Right: allow leading spaces (non-capturing), then capture text
-                        // For _expression compatibility, use " *(.+?)" format (leading spaces, then capture)
-                        Some('>') => r" *(.+?)".to_string(),
-                        // Center: allow spaces on both sides (non-capturing), capture text in middle
-                        Some('^') => r"(?:\s*)([^\{\}\s]+(?:\s+[^\{\}\s]+)*?)(?:\s*)".to_string(),
-                        _ => r"[^\{\}]+?".to_string(),
-                    }
-                } else {
-                    // For empty {} fields, match any characters including newlines (non-greedy).
-                    // Delimited default strings may match empty (issue #83); otherwise require
-                    // at least one character like the original parse library.
-                    if allow_empty_delimited {
-                        r".*?".to_string()
-                    } else {
-                        r".+?".to_string()
-                    }
-                }
+        match &self.field_type {
+            FieldType::String | FieldType::Multiline | FieldType::IndentBlock => {
+                string::pattern(self, next_field_is_greedy, allow_empty_delimited)
             }
-            FieldType::Multiline | FieldType::IndentBlock => {
-                // Same layout as strings, but ``.`` cannot span newlines in Rust regex; use ``[\s\S]``.
-                // ``IndentBlock`` (:blk) uses the same fragment as ``Multiline``; dedent is applied after capture.
-                const ML: &str = "[\\s\\S]";
-                if let Some(prec) = self.precision {
-                    if let Some(align) = self.alignment {
-                        let fill_ch = self.fill.unwrap_or(' ');
-                        let fill_escaped = regex::escape(&fill_ch.to_string());
-                        let width_equals_precision = self.width == Some(prec);
-                        match align {
-                            '<' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", ML, prec)
-                                } else {
-                                    format!("{}{{{}}}(?:{}*)", ML, prec, fill_escaped)
-                                }
-                            }
-                            '>' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", ML, prec)
-                                } else {
-                                    format!("(?:{}*?){}{{{}}}", fill_escaped, ML, prec)
-                                }
-                            }
-                            '^' => {
-                                if width_equals_precision {
-                                    format!("{}{{{}}}", ML, prec)
-                                } else {
-                                    format!(
-                                        "(?:{}*?){}{{{}}}(?:{}*)",
-                                        fill_escaped, ML, prec, fill_escaped
-                                    )
-                                }
-                            }
-                            _ => format!("{}{{{}}}", ML, prec),
-                        }
-                    } else {
-                        format!("{}{{{}}}", ML, prec)
-                    }
-                } else if let Some(width) = self.width {
-                    match next_field_is_greedy {
-                        Some(false) => format!("{}{{{}}}", ML, width),
-                        _ => format!("{}{{{},}}", ML, width),
-                    }
-                } else if self.alignment.is_some() {
-                    match self.alignment {
-                        Some('<') => format!("({}+?)(?:\\s*)", ML),
-                        Some('>') => format!(" *({}+?)", ML),
-                        Some('^') => format!("(?:\\s*)({}+?)(?:\\s*)", ML),
-                        _ => format!("{}+?", ML),
-                    }
-                } else {
-                    r"[\s\S]+?".to_string()
-                }
-            }
-            FieldType::Integer => {
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => r"[+-]?", // Default: allow optional + or -
-                    })
-                    .unwrap_or(r"[+-]?"); // Default: allow optional + or -
-
-                // Handle fill character with alignment (e.g., {:x=5d})
-                // For '=' alignment, fill goes between sign and digits
-                // Pattern should match: [sign][fill*][digits]
-                let (fill_prefix, fill_suffix) =
-                    if let (Some(fill_ch), Some('=')) = (self.fill, self.alignment) {
-                        // For '=' alignment with fill, match fill characters between sign and number
-                        let fill_escaped = regex::escape(&fill_ch.to_string());
-                        (format!("{}*", fill_escaped), String::new())
-                    } else {
-                        (String::new(), String::new())
-                    };
-
-                let width_precision =
-                    int_width_precision_bounds(self.width, self.precision, self.zero_pad);
-
-                let base_pattern = if self.zero_pad {
-                    if let Some((lo, hi)) = width_precision {
-                        // Both width and precision: bounded digit count (formatparse#82 / parse#107).
-                        format!(
-                            "{}{}{}[0-9]{{{},{}}}",
-                            sign, fill_prefix, fill_suffix, lo, hi
-                        )
-                    } else if let Some(width) = self.width {
-                        // Zero-padded: if width is specified, match 1 to width digits
-                        // This allows unpadded values (e.g., '9' for {c:02d}) but rejects values exceeding width
-                        format!("{}{}{}[0-9]{{1,{}}}", sign, fill_prefix, fill_suffix, width)
-                    } else {
-                        format!("{}{}{}[0-9]+", sign, fill_prefix, fill_suffix)
-                    }
-                } else if let Some((lo, hi)) = width_precision {
-                    let q_hex = format!("[0-9a-fA-F]{{{},{}}}", lo, hi);
-                    let q_oct = format!("[0-7]{{{},{}}}", lo, hi);
-                    let q_bin = format!("[01]{{{},{}}}", lo, hi);
-                    let q_dec = format!("[0-9]{{{},{}}}", lo, hi);
-                    match self.original_type_char {
-                        Some('x') | Some('X') => {
-                            // Hex: bounded digit run after optional 0x (parse#107 / #82).
-                            format!(
-                                "{}{}{}(?:0[xX]{}|{})",
-                                sign, fill_prefix, fill_suffix, q_hex, q_hex
-                            )
-                        }
-                        Some('o') => {
-                            format!(
-                                "{}{}{}(?:0[oO]{}|{})",
-                                sign, fill_prefix, fill_suffix, q_oct, q_oct
-                            )
-                        }
-                        Some('b') => {
-                            format!(
-                                "{}{}{}(?:0[bB]{}|{})",
-                                sign, fill_prefix, fill_suffix, q_bin, q_bin
-                            )
-                        }
-                        _ => {
-                            // Decimal `d` / `i`: optional leading whitespace before digits (#81);
-                            // each radix branch uses the same inclusive min/max digit count.
-                            format!(
-                                "{}{}{}(?:0[xX]{}|0[oO]{}|0[bB]{}|\\s*{})",
-                                sign, fill_prefix, fill_suffix, q_hex, q_oct, q_bin, q_dec
-                            )
-                        }
-                    }
-                } else {
-                    // Check original type to determine what digits to match
-                    match self.original_type_char {
-                        Some('x') | Some('X') => {
-                            // Hex: match hex digits with or without 0x prefix
-                            format!(
-                                "{}{}{}(?:0[xX][0-9a-fA-F]+|[0-9a-fA-F]+)",
-                                sign, fill_prefix, fill_suffix
-                            )
-                        }
-                        Some('o') => {
-                            // Octal: match octal digits with or without 0o prefix
-                            format!(
-                                "{}{}{}(?:0[oO][0-7]+|[0-7]+)",
-                                sign, fill_prefix, fill_suffix
-                            )
-                        }
-                        Some('b') => {
-                            // Binary: match binary digits with or without 0b prefix
-                            format!("{}{}{}(?:0[bB][01]+|[01]+)", sign, fill_prefix, fill_suffix)
-                        }
-                        _ => {
-                            // Decimal: optional leading whitespace before digits (parse#133 / #81)
-                            // so values like "    0" match like str.format padding, without widening
-                            // the 0x/0o/0b branches.
-                            format!(
-                                "{}{}{}(?:0[xX][0-9a-fA-F]+|0[oO][0-7]+|0[bB][01]+|\\s*[0-9]+)",
-                                sign, fill_prefix, fill_suffix
-                            )
-                        }
-                    }
-                };
-
-                wrap_field_lookbehind(self, base_pattern)
-            }
-            FieldType::Float => {
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => r"[+-]?", // Default: allow optional + or -
-                    })
-                    .unwrap_or(r"[+-]?"); // Default: allow optional + or -
-
-                // For floats, precision affects how we match
-                // Width is mainly for formatting, but we need to handle it in parsing
-                // When width is specified, there may be leading/trailing spaces
-                let inner = if let Some(prec) = self.precision {
-                    // Precision specified — digits after the decimal when a decimal point appears.
-                    // For prec == 0, str.format often emits no '.' (e.g. "{:02.0f}" → "20"); accept
-                    // integer-looking text as well as "12." / ".5" forms (issue #84 / parse#159).
-                    if prec == 0 {
-                        let width_prefix = if self.width.is_some() { r"\s*" } else { "" };
-                        format!(
-                            r"{}{}(?:\d+(?:\.0*)?|\d*\.\d{{{}}}|\.\d{{{}}})(?:[eE][+-]?\d+)?",
-                            width_prefix, sign, prec, prec
-                        )
-                    } else if self.width.is_some() {
-                        // Width specified - allow optional leading spaces
-                        format!(
-                            r"\s*{}(?:\d*\.\d{{{}}}|\.\d{{{}}})(?:[eE][+-]?\d+)?",
-                            sign, prec, prec
-                        )
-                    } else {
-                        format!(
-                            r"{}(?:\d*\.\d{{{}}}|\.\d{{{}}})(?:[eE][+-]?\d+)?",
-                            sign, prec, prec
-                        )
-                    }
-                } else {
-                    // Float must have a decimal point (not just an integer)
-                    // Allow: 12.34, .34, 12., or scientific notation with decimal
-                    format!(r"{}(?:\d+\.\d+|\.\d+|\d+\.)(?:[eE][+-]?\d+)?", sign)
-                };
-                wrap_field_lookbehind(self, inner)
-            }
-            FieldType::Letters => r"[a-zA-Z]+".to_string(),
-            FieldType::Word => r"\w+".to_string(),
-            FieldType::NonLetters => r"[^a-zA-Z]+".to_string(),
-            FieldType::NonWhitespace => r"\S+".to_string(),
-            FieldType::NonDigits => r"[^0-9]+".to_string(),
-            FieldType::NumberWithThousands => {
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => r"[+-]?", // Default: allow optional + or -
-                    })
-                    .unwrap_or(r"[+-]?"); // Default: allow optional + or -
-                                          // Match numbers with thousands separators (comma or dot)
-                                          // Pattern: either number with valid thousands separators (1,234,567 or 1.234.567)
-                                          // or plain number without separators
-                                          // The regex matches the pattern, validation happens in conversion
-                format!(r"{}(?:\d{{1,3}}(?:[.,]\d{{3}})*|\d+)", sign)
-            }
-            FieldType::Scientific => {
-                // Scientific notation: matches floats with e/E exponent, or nan/inf
-                // Pattern matches original parse library exactly: \d*\.\d+[eE][-+]?\d+|nan|NAN|[-+]?inf|[-+]?INF
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => "-?",
-                    })
-                    .unwrap_or("-?");
-                // Sign applies to numeric part; nan/inf have their own optional signs in the pattern
-                format!(r"{}\d*\.\d+[eE][-+]?\d+|nan|NAN|[-+]?inf|[-+]?INF", sign)
-            }
-            FieldType::GeneralNumber => {
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => "-?",
-                    })
-                    .unwrap_or("-?");
-                // General number: can be int or float or scientific, or nan/inf
-                format!(
-                    r"{}(?:\d+\.\d+|\.\d+|\d+\.|\d+)(?:[eE][+-]?\d+)?|nan|NAN|[-+]?inf|[-+]?INF",
-                    sign
-                )
-            }
-            FieldType::Percentage => {
-                let sign = self
-                    .sign
-                    .as_ref()
-                    .map(|s| match s {
-                        '+' => r"\+?",
-                        '-' => "-?",
-                        ' ' => r"[- ]?",
-                        _ => "-?",
-                    })
-                    .unwrap_or("-?");
-                // Percentage: number followed by %
-                format!(r"{}(?:\d+\.\d+|\.\d+|\d+)%", sign)
-            }
-            FieldType::DateTimeISO => {
-                // ISO 8601 format: YYYY-MM-DD, YYYY-MM-DDTHH:MM, YYYY-MM-DDTHH:MM:SS, etc.
-                // Supports various separators and timezone formats (with optional space before timezone)
-                r"\d{4}-\d{2}-\d{2}(?:[T ]\d{2}:\d{2}(?::\d{2}(?:\.\d+)?)?)?(?:\s*[Zz]|\s*[+-]\d{2}:?\d{2}|\s*[+-]\d{4})?".to_string()
-            }
-            FieldType::DateTimeRFC2822 => {
-                // RFC2822: Mon, 21 Nov 2011 10:21:36 +1000 or +10:00 (optional weekday)
-                r"(?:(?:Mon|Tue|Wed|Thu|Fri|Sat|Sun),\s+)?\d{1,2}\s+(?:Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\s+\d{4}\s+\d{2}:\d{2}:\d{2}\s+[+-]\d{2}:?\d{2,4}".to_string()
-            }
-            FieldType::DateTimeGlobal => {
-                // Global format: 21/11/2011 10:21:36 AM +1000 or 21-Nov-2011 10:21:36 AM +1:00
-                r"\d{1,2}[-/](?:\d{1,2}|Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec|January|February|March|April|June|July|August|September|October|November|December)[-/]\d{4}(?:\s+\d{1,2}:\d{2}(?::\d{2})?(?:\s+[AP]M)?(?:\s+[+-]\d{1,2}:?\d{2,4})?)?".to_string()
-            }
-            FieldType::DateTimeUS => {
-                // US format: 11/21/2011 10:21:36 AM +1000 or 11-Nov-2011 10:21:36 AM +1000 or Nov-21-2011 10:21:36 AM +1000
-                r"(?:\d{1,2}|Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec|January|February|March|April|June|July|August|September|October|November|December)[-/]\d{1,2}[-/]\d{4}(?:\s+\d{1,2}:\d{2}(?::\d{2})?(?:\s+[AP]M)?(?:\s+[+-]\d{2}:?\d{2,4})?)?".to_string()
-            }
-            FieldType::DateTimeCtime => {
-                // ctime format: Mon Nov 21 10:21:36 2011
-                r"(?:Mon|Tue|Wed|Thu|Fri|Sat|Sun)\s+(?:Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)\s+\d{1,2}\s+\d{2}:\d{2}:\d{2}\s+\d{4}".to_string()
-            }
-            FieldType::DateTimeHTTP => {
-                // HTTP log format: 21/Nov/2011:00:07:11 +0000
-                r"\d{2}/[A-Za-z]{3}/\d{4}:\d{2}:\d{2}:\d{2}\s+[+-]\d{2}:?\d{2,4}".to_string()
-            }
-            FieldType::DateTimeTime => {
-                // Time format: 10:21:36 PM -5:30
-                r"\d{1,2}:\d{2}(?::\d{2})?(?:\s+[AP]M)?(?:\s+[+-]\d{1,2}:?\d{2,4})?".to_string()
-            }
-            FieldType::DateTimeSystem => {
-                // Linux system log format: Nov 21 10:21:36
-                r"[A-Za-z]{3}\s+\d{1,2}\s+\d{2}:\d{2}:\d{2}".to_string()
-            }
-            FieldType::DateTimeStrftime => {
-                // Convert strftime format to regex pattern
-                if let Some(ref fmt) = self.strftime_format {
-                    strftime_to_regex(fmt)
-                } else {
-                    r".+?".to_string()
-                }
-            }
-            FieldType::BracedContent => {
-                // Placeholder: named patterns use special regex in the PyO3 pattern compiler.
-                r".*?".to_string()
-            }
-            FieldType::Nested => self
-                .nested_regex_body
-                .as_deref()
-                .unwrap_or(r"[\s\S]*?")
-                .to_string(),
-            FieldType::Boolean => {
-                "true|false|True|False|TRUE|FALSE|1|0|yes|no|Yes|No|YES|NO|on|off|On|Off|ON|OFF"
-                    .to_string()
-            }
-            FieldType::Custom(name) => {
-                custom_patterns
-                    .get(name)
-                    .cloned()
-                    .unwrap_or_else(|| r"\S+".to_string()) // Default to non-whitespace for custom types without patterns
-            }
-        };
-
-        base_pattern
+            FieldType::Integer
+            | FieldType::Float
+            | FieldType::Letters
+            | FieldType::Word
+            | FieldType::NonLetters
+            | FieldType::NonWhitespace
+            | FieldType::NonDigits
+            | FieldType::NumberWithThousands
+            | FieldType::Scientific
+            | FieldType::GeneralNumber
+            | FieldType::Percentage => numeric::pattern(self),
+            FieldType::DateTimeISO
+            | FieldType::DateTimeRFC2822
+            | FieldType::DateTimeGlobal
+            | FieldType::DateTimeUS
+            | FieldType::DateTimeCtime
+            | FieldType::DateTimeHTTP
+            | FieldType::DateTimeTime
+            | FieldType::DateTimeSystem
+            | FieldType::DateTimeStrftime => datetime_fragment::pattern(self),
+            FieldType::BracedContent
+            | FieldType::Nested
+            | FieldType::Boolean
+            | FieldType::Custom(_) => custom::pattern(self, custom_patterns),
+        }
     }
 }
-
