+++ formatparse-pyo3/src/parser/matching/capture.rs
@@ -1,15 +1,7 @@
-use crate::error;
-use crate::match_rs::{Match, MatchInit};
-use crate::parser::format_parser::FormatParser;
-use crate::parser::raw_match::{RawMatchData, RawValue};
-use crate::result::ParseResult;
-use fancy_regex::{Captures, Regex};
+use fancy_regex::Captures;
 use formatparse_core::{count_capturing_groups, FieldSpec, FieldType};
 use pyo3::prelude::*;
-use pyo3::types::PyDict;
-use pyo3::IntoPyObjectExt;
 use std::collections::HashMap;
-use std::sync::Arc;
 
 use super::custom_type::validate_custom_type_pattern;
 
@@ -111,4 +103,3 @@ pub(crate) fn strftime_merge_leader_per_field(
     }
     out
 }
-
