+++ formatparse-pyo3/src/parser/matching/custom_type.rs
@@ -1,15 +1,7 @@
 use crate::error;
-use crate::match_rs::{Match, MatchInit};
-use crate::parser::format_parser::FormatParser;
-use crate::parser::raw_match::{RawMatchData, RawValue};
-use crate::result::ParseResult;
-use fancy_regex::{Captures, Regex};
 use formatparse_core::{count_capturing_groups, FieldSpec, FieldType};
 use pyo3::prelude::*;
-use pyo3::types::PyDict;
-use pyo3::IntoPyObjectExt;
 use std::collections::HashMap;
-use std::sync::Arc;
 
 pub fn validate_custom_type_pattern(
     field_spec: &FieldSpec,
@@ -74,4 +66,3 @@ pub fn validate_custom_type_pattern(
 
     Ok(pattern_groups)
 }
-
