+++ tests/test_findall.py
@@ -3,18 +3,14 @@ from formatparse import findall
 
 
 def test_findall():
-    s = "".join(
-        r.fixed[0] for r in findall(">{}<", "<p>some <b>bold</b> text</p>")
-    )
+    s = "".join(r.fixed[0] for r in findall(">{}<", "<p>some <b>bold</b> text</p>"))
     assert s == "some bold text"
 
 
 def test_no_evaluate_result():
     s = "".join(
         m.evaluate_result().fixed[0]
-        for m in findall(
-            ">{}<", "<p>some <b>bold</b> text</p>", evaluate_result=False
-        )
+        for m in findall(">{}<", "<p>some <b>bold</b> text</p>", evaluate_result=False)
     )
     assert s == "some bold text"
 
@@ -23,9 +19,7 @@ def test_case_sensitivity():
     results = [r.fixed[0] for r in findall("x({})x", "X(hi)X")]
     assert results == ["hi"]
 
-    results = [
-        r.fixed[0] for r in findall("x({})x", "X(hi)X", case_sensitive=True)
-    ]
+    results = [r.fixed[0] for r in findall("x({})x", "X(hi)X", case_sensitive=True)]
     assert results == []
